Update index.js
Modified function downloadAndFlash to handle image offset.
Image offset should be added to config.toml inside the section of the corresponding app:
Example of (partial) config.toml, with offset 0x10000:

supported_apps = ["example_app"]
...
[example_app]
offset = "0x10000"

Also, added an alert if the image file is not found, because otherwise there is no warning to the user
(temporary solution, needs to be revised for better error handling).